requirements.txt
alembic>=1.12.0
asyncpg>=0.28.0
psycopg>=3.1.10
psycopg-binary>=3.1.10
passlib[bcrypt]>=1.7.4
sqlalchemy>=2.0.20

fastapi>=0.115.0
pydantic>=2.3.0
pydantic-settings>=2.0.3
uvicorn>=0.23.2

bcrypt>=4.0.1
pytest>=8.3.3
python-jose>=3.3.0
